Retry on login failure
Prior to this commit, the sync requests (and other requests) would not
happen in the sceanario that the login request fails.

With the changes on this commit: if the OIDC server is down, it will
retry logging in up to 3 times. And once it succeeds then it will run
the other requests (including sync).