Improves metric collection
This tries to separate metrics to avoid scenarios where the response
time is very fast but it's only because the backend server is down.

For example, sometimes the sync worker will get knocked offline (e.g.
OOM), and while it's down nginx will quickly respond with either a 500
or a 503.

While the fast response time is technically accurate, a fast response
time on a 5xx error just skews the overall response time stats.