Creates test data used by load test
This commit adds support to run the following command:

poetry run python locust-setup-test-data.py --host https://example.com
--rooms=1 --messages=1 --reactions=1

The script logs in as each user in the setup-users.csv list. And then
each user will create the specified number of rooms, messages and
reactions. The number of rooms, messages, and reactions is configurable
by the command line args above.

It should be noted that the room memberships are populated via
Audiences, which handles room memberships via SCIM (so that entire
groups of users can be added to a room and so that room memberships stay
up to date as the SCIM group changes over time).
https://github.com/powerhome/audiences

The goal is to setup test data so that when the load test happens, the
test user performing the load test are syncing existing rooms, messages,
and reaction events. Part of the reason sync requests are heavy is
because they include m.room.member events and thus it is crucial to
include rooms with many members in a load test.