Load test for sync (with OIDC auth)
This commit adds support to run the following command:

poetry run locust -f oidc-login-sync-test.py

When you run this command, you can run the Locust web ui by visiting
http://0.0.0.0:8089. And from there, you can run load tests that
excercise the login flow that OIDC clients typically use on a Matrix
chat server.

For example, the test users will:

1. Login via OIDC (this depends on an OIDC server)
2. Fetch the current user's avatar and displayname from the homeserver.
3. And most importantly the client will perform an initial sync.

When you run the test via the Locustg web UI, there are options to
select the host, the number of users, the spawn rate, and the sync type.

The sync type currently has two options: standard and lazy-loading.

The goal of this commit is to measure and analyze the difference in load
times and server-side performance when changing the sync-type option.